增加 service 写法支持，增加编译调试模块分离
