--- packages/mip-cli/lib/builder/index.js
@@ -1,8 +1,0 @@
-/**
- * @file builder.js
- * @author clark-t (clarktanglei@163.com)
- */
-
-const WebpackBuilder = require('./webpack')
-
-module.exports = WebpackBuilder
