+++ packages/mip-cli/package.json
@@ -83,7 +83,8 @@
     "webpack": "^4.8.3",
     "webpack-dev-middleware": "^3.1.3",
     "webpack-sources": "^1.1.0",
-    "workbox-build": "3.2.0"
+    "workbox-build": "3.2.0",
+    "wrapper-webpack-plugin": "^2.0.0"
   },
   "engines": {
     "node": ">=7.6"
